move opentracer fixtures to conftest.py
